Lint fix; update cast table to include unspecified values

+++ onnx/reference/ops/op_cast.py
@@ -29,9 +29,7 @@ def cast_to(
         return onnx.numpy_helper.saturating_cast(x, dtype)
 
     if to == onnx.TensorProto.FLOAT8E8M0:
-        return onnx.numpy_helper.to_float8e8m0(x, saturate, round_mode).astype(
-            dtype
-        )
+        return onnx.numpy_helper.to_float8e8m0(x, saturate, round_mode).astype(dtype)
 
     return x.astype(dtype)
